requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
torch
transformers
numpy
nest-asyncio
